refactor(solver): group multi-file backends into subdirs
Each solver kind with more than one file now lives in its own
sub-package under ``src/qkan/solver/``:

- ``solver/cute/``   — cute.py, cute_activation.py, cute_linear.py, cute_ops.py
- ``solver/cutile/`` — cutile.py, cutile_activation.py, cutile_activations.py,
                      cutile_linear.py, cutile_ops.py
- ``solver/flash/``  — flash.py, triton_activation.py, triton_activations.py,
                      triton_linear.py, fused_ops.py

Single-file backends (exact, qml, qiskit, cutn, cudaq) stay flat.

Top-level kernel modules (``qkan.cute_ops``, ``qkan.cutile_ops``,
``qkan.fused_ops``) moved into their respective solver sub-packages.
Internal imports updated; ``parents[2]`` → ``parents[4]`` in
``cute_ops`` for the now-deeper path-to-csrc lookup.

Public API in ``qkan/__init__.py`` and the ``solver`` re-exports
(``flash_exact_solver``, ``cute_exact_solver``, etc.) are unchanged.
