+++ src/qkan/solver/cutile/cutile_linear.py
