+++ src/qkan/solver/flash/fused_ops.py
