+++ src/qkan/solver/flash/triton_activations.py
