+++ src/qkan/solver/flash/triton_linear.py
